[FEAT] agrego el componente perfil
