+++ src/styles/perfil/css/Perfil.css
@@ -1,18 +1,16 @@
-html{
+.App#perfil{
     background-image:  url(../img/perfil/fondoPerfil.jpg);
 }
 .nav-container .hamburger-lines .line {
     background: var(--hamburger_scrolHandle_Mapas_Perfil);
 }
-header{
-    overflow: hidden;
-}
-.sistema_solar {
+/* .sistema_solar {
     position: relative;
     grid-row: 2;
     top: -150px;
     left: 33vw;
-}
+} */
+
 .cuote{
     grid-column: 1/4;
     grid-row: 3;
@@ -37,8 +35,8 @@ header{
     grid-template-columns: repeat(2, 1fr);
 }
 .top_asistencia, .comidas{
-    padding-left: 25px;
-    padding-right: 25px;
+    padding: 25px;
+    padding-top: 0;
 }
 .comidas{
     margin-top: 30px;
@@ -377,4 +375,78 @@ iframe{
 }
 #logo img{
     height: 100px;
+}
+
+
+@media (min-width: 1024px){
+    #credencial {
+        grid-column: 1/3;
+    }
+    /* PERFIL */
+    .perfil header{
+        height: 100vh;
+    }
+    .perfil .sistema_solar {
+        top: -370px;
+    }
+    .perfil .sol {
+        background: var(--sol_planetas);
+        border-radius: 100%;
+        width: 450px;
+        height: 450px;
+        z-index: -5;
+    }
+    .perfil .orbit_1 {
+        width: 680px;
+        height: 680px;
+        margin: -110px auto;
+    }
+    .perfil .orbit_2 {
+        width: 880px;
+        height: 880px;
+        margin: -210px auto;
+    }
+    .perfil .orbit_3 {
+        width: 1080px;
+        height: 1080px;
+        margin: -310px auto;
+    }
+    .perfil .planet {
+        width: 75px;
+        height: 75px;
+    }
+    .perfil .orbit_1 .planet {
+        margin: 8%;
+    }
+    .cuote {
+        grid-row: 4;
+        padding-top: 15%;
+    }
+    .coheteconteiner{
+        display: none;
+    }
+    .top_asistencia .astronauta_de_caeii {
+        width: auto;
+    }
+    /* #mapas, #cronograma{
+        height: 145vh;
+        margin: 0;
+    } */
+    .actividad, th {
+        font-size: 1.7rem;
+        padding: 30px;
+        padding-left: 0;
+        padding-right: 0;
+    }
+    iframe{
+        width: 40vw;
+        height: 50vh;
+        margin-top: 40px;
+    }
+    #logo{
+        grid-column: 1/3;
+    }
+    #logo img {
+        height: 150px;
+    }
 }
