--- src/styles/perfil/css/fuentes.css
@@ -1,24 +1,0 @@
-@font-face {
-    /* fuente general */
-    font-family: 'Montserrat-Regular';
-    src: url(../fonts/Montserrat-Regular.ttf) format('truetype');
-}
-@font-face {
-    /* fuente de los h1 */
-    font-family: 'Montserrat-Italic';
-    src: url(../fonts/Montserrat-Italic.ttf) format('truetype');
-}
-@font-face {
-    /* fuente de los h2 */
-    font-family: 'Montserrat-Black';
-    src: url(../fonts/Montserrat-Black.ttf) format('truetype');
-}
-@font-face { 
-    /* fuente de los links del menu de home */
-     font-family: 'Montserrat-MediumItalic';
-    src: url(../fonts/Montserrat-MediumItalic.ttf) format('truetype');
-}
-@font-face {
-    font-family: 'Montserrat-LightItalic';
-    src: url(../fonts/Montserrat-LightItalic.ttf) format('truetype');
-}
