--- src/styles/perfil/css/generales/desktop.css
@@ -1,253 +1,0 @@
-@media (min-width: 1024px) {
-    header{
-        grid-template-rows: 150px 200px;
-    }
-    .movile_navbar{
-        display: none;
-    }
-    .desktop_navbar{
-        display: block;
-        height: 110px;
-    }
-    main {
-        padding: 40px;
-        padding-top: 0;
-        display: grid;
-        grid-template-columns: repeat(2, 1fr);
-        column-gap: 100px
-    }
-
-    /* HOME */
-    .sistema_solar{
-        top: 620px;
-    }
-    .fondo {
-        width: 100vw;
-        height: 100vw;
-        margin: -50vw;
-    }
-    li{
-        margin-top: 0;
-        height: 75px;
-    }
-    #menu a {
-        font-size: 1.8rem;
-    }
-    #uno, #seis{
-        grid-column: 1/9;
-    }
-    #dos, #cinco{
-        grid-column: 3/8;
-    }
-    #tres, #cuatro{
-        grid-column: 2/9;
-    }
-    #registrarse{
-        margin-top: 0;
-    }
-
-    #estaciones, #sponsors{
-        grid-column: 1/3;
-    }
-
-    section h2{
-        height: 110px;
-        font-size: 2rem;
-    }
-
-    #estaciones article{
-        grid-template-rows: 100% !important;
-        grid-template-columns: 40% 20% 40%;
-    }
-    #estaciones article h3, #estaciones .derecha p{
-        grid-column: 1/2;
-    }
-    #estaciones article h3{
-        font-size: 1.5rem;
-        justify-content: center;
-    }
-    #estaciones article .img, #estaciones .derecha .img{
-        grid-column: 2/3;
-    }
-    #estaciones article p, #estaciones .derecha h3{
-        grid-column: 3/4;
-    }
-    #estaciones article p{
-        font-size: 1.1rem;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-    }
-    #estaciones .derecha p, #estaciones .derecha h3{
-        grid-row: 1;
-    }
-    #estaciones .imgOrbit {
-        height: 150px;
-        width: 150px;
-        left: 0;
-    }
-    #estaciones .leermas{
-        margin: 0;
-        position: absolute;
-        top: 1015px;
-        left: 80vw;
-    }
-    #estaciones .desktop{
-        display: flex;
-    }
-    #estaciones .movile{
-        display: none;
-    }
-    #estaciones .derecha *{
-        text-align: left;
-    }
-
-    /* PERFIL */
-    .perfil header{
-        height: 100vh;
-    }
-    .perfil .sistema_solar {
-        top: -370px;
-    }
-    .perfil .sol {
-        background: var(--sol_planetas);
-        border-radius: 100%;
-        width: 450px;
-        height: 450px;
-        z-index: -5;
-    }
-    .perfil .orbit_1 {
-        width: 680px;
-        height: 680px;
-        margin: -110px auto;
-    }
-    .perfil .orbit_2 {
-        width: 880px;
-        height: 880px;
-        margin: -210px auto;
-    }
-    .perfil .orbit_3 {
-        width: 1080px;
-        height: 1080px;
-        margin: -310px auto;
-    }
-    .perfil .planet {
-        width: 75px;
-        height: 75px;
-    }
-    .perfil .orbit_1 .planet {
-        margin: 8%;
-    }
-    .cuote {
-        grid-row: 4;
-        padding-top: 15%;
-    }
-    .coheteconteiner{
-        display: none;
-    }
-    #credencial{
-        grid-column: 1/3;
-        margin-bottom: 100px;
-        padding-bottom: 5%;
-        display: grid;
-        grid-template-columns: 70% 30%;
-    }
-    #credencial h2{
-        grid-column: 1/3;
-    }
-    #credencial .top_asistencia{
-        grid-column: 1/2;
-        padding-right: 0;
-        padding-left: 0;
-    }
-    #credencial .top_asistencia h3 {
-        font-size: 1.8rem;
-    }
-    #credencial .comidas{
-        grid-column: 2/3;
-        justify-content: center;
-        align-items: stretch;
-        margin-top: 0;
-    }
-    #credencial .almuerzos .iconos, #credencial .coffees .iconos {
-        margin-left: 0;
-        margin-top: 20px;
-        margin-bottom: 30px;
-        display: flex;
-    }
-    #credencial .comidas h3 {
-        font-size: 1.6rem;
-    }
-    #credencial .top_asistencia .astronauta_de_caeii {
-        width: 25vw;
-        margin-left: 60px;
-        height: auto;
-    }
-    #credencial .top_asistencia .circulo_qr {
-        height: 25vw;
-        width: 25vw;
-    }
-    #credencial .asistencia{
-        margin-left: -20px;
-        width: auto;
-    }
-    #credencial .modal {
-        padding-top: 35vh;
-    }
-    #credencial .modal-content img {
-        width: 30vh;
-        height: 30vh;
-    }
-    #mapas, #cronograma{
-        height: 145vh;
-        margin: 0;
-    }
-    .actividad, th {
-        font-size: 1.7rem;
-        padding: 30px;
-        padding-left: 0;
-        padding-right: 0;
-    }
-    iframe{
-        width: 40vw;
-        height: 50vh;
-        margin-top: 40px;
-    }
-    #logo{
-        grid-column: 1/3;
-    }
-    #logo img {
-        height: 150px;
-    }
-
-    /* INFO GENERAL */
-    .info_card h3{
-        font-size: 1.9rem;
-    }
-    .info_card p{
-        font-size: 1.5rem;
-    }
-    .info_card .info_items li{
-        font-size: 1.7rem;
-    }
-}
-@media (min-width: 1300px) {
-    .imgOrbit{
-        left: -2vw;
-    }
-    .derecha .imgOrbit {
-        left: 2vw;
-    }
-    #credencial .comidas .iconos img {
-        height: 60px;
-    }
-    #credencial .asistencia{
-        margin-left: -80px;
-    }
-}
-@media only screen and (max-width: 1260px){
-    #cinco .line_container {
-        margin-left: 15%;
-        margin-right: 15%;
-    }
-}
